+++ xdp-ebpf/src/lib.rs
@@ -17,15 +17,10 @@
 #![no_std]
 
 #[cfg(all(target_os = "linux", not(target_arch = "bpf")))]
+pub static AGAVE_XDP_EBPF_PROGRAM: &[u8] =
+    aya::include_bytes_aligned!(concat!(env!("CARGO_MANIFEST_DIR"), "/agave-xdp-prog"));
+
+#[cfg(all(target_os = "linux", not(target_arch = "bpf")))]
 #[unsafe(no_mangle)]
-pub static AGAVE_XDP_EBPF_PROGRAM: [u8; aya::include_bytes_aligned!(concat!(
-    env!("CARGO_MANIFEST_DIR"),
-    "/agave-xdp-prog"
-))
-.len()] = unsafe {
-    core::ptr::read(
-        aya::include_bytes_aligned!(concat!(env!("CARGO_MANIFEST_DIR"), "/agave-xdp-prog"))
-            .as_ptr()
-            .cast(),
-    )
-};
+pub static AGAVE_XDP_EBPF_PROGRAM_BYTES: [u8; AGAVE_XDP_EBPF_PROGRAM.len()] =
+    unsafe { core::ptr::read(AGAVE_XDP_EBPF_PROGRAM.as_ptr().cast()) };