+++ CHANGELOG.md
@@ -30,6 +30,8 @@ Release channels have their own copy of this changelog:
   * `--dev-halt-at-slot`
 #### Deprecations
 * Using `mmap` for `--accounts-db-access-storages-method` is now deprecated.
+#### Changes
+* Implement XDP Dispatcher
 
 ## 3.1.0
 ### RPC